# Conflicts:
#	features/discovery/home_component/src/main/java/com/tokopedia/home_component/viewholders/ReminderWidgetViewHolder.kt
#	libraries/applink/src/main/java/com/tokopedia/applink/DeeplinkDFMapper.kt

+++ features/merchant/product_add_edit/src/main/java/com/tokopedia/product/addedit/preview/data/source/api/response/Variant.kt
@@ -1,6 +1,7 @@
 package com.tokopedia.product.addedit.preview.data.source.api.response
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigInteger
 
 data class Variant(
         @SerializedName("products")
@@ -21,7 +22,7 @@ data class ProductVariant(
         @SerializedName("isPrimary")
         val isPrimary: Boolean,
         @SerializedName("price")
-        val price: Long,
+        val price: BigInteger,
         @SerializedName("sku")
         val sku: String,
         @SerializedName("stock")
